requirements.txt
black==21.12b0
click==8.0.3
colorama==0.4.4
imutils==0.5.4
mypy-extensions==0.4.3
numpy==1.21.5
opencv-contrib-python==4.5.5.62
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
tomli==1.2.3
typing_extensions==4.0.1
